backend/requirements.txt
fastapi
uvicorn
pdfplumber
google-generativeai
python-dotenv
python-multipart
pydantic
